+++ src/app/components/PictureDocket.js
