+++ src/app/components/views/index.js
@@ -2,5 +2,5 @@
 
 // return
 export { default as About } from "./About"
-export { default as NotFound } from "./NotFound"
 export { default as SignIn } from "./SignIn"
+export { default as NotFound } from "./NotFound"
