+++ src/app/containers/Composer/index.js
@@ -13,18 +13,18 @@ const titlePlaceholder = {
 }
 
 // return
-export default class extends React.Component {
-	render() {
-		return(
-			<div>
-				<HeaderEditor
-  				pageTitle={ titlePlaceholder.title }
-  				pageSubtitle={ titlePlaceholder.subtitle }
-  			/>
-				<ContentEditor
-					ref={ input => { this.contentEditor = input } }
-				/>
-			</div>
-		)
-	}
+// this is a pure function, but I left it in /containers folder
+// because it gathers all sub-containers conveniently for export
+export default props => {
+	return(
+		<div>
+			<HeaderEditor
+				pageTitle={ titlePlaceholder.title }
+				pageSubtitle={ titlePlaceholder.subtitle }
+			/>
+			<ContentEditor
+				ref={ input => { this.contentEditor = input } }
+			/>
+		</div>
+	)
 }
